refactor + feat: move CreateCast + get it pretty much working minus actually posting
